Modified CMakeLists.txt so that it uses the FindVRPN and Findquatlib tools to locate these libraries.  Added these Find commands to the cmake directory.  Fixed a signed/unsigned mismatch in the code to avoid compiler warnings.  It now compiles on Windows except for not finding PQP.

+++ modelmanager.cpp
@@ -86,7 +86,7 @@ int ModelManager::addObjectType(int srcIndex, double scale) {
     transformPD->Update();
 
 
-    int idNum = models.size();
+    size_t idNum = models.size();
     // fix potential out of space problems as they come up
     if (idNum >= models.capacity()) {
         models.reserve(2*idNum);
